Rename `modal-class` => `modal-manager`
