+++ src/medmain-modal/modal-manager.js
