+++ src/medmain-modal/stories/index.js
@@ -1,3 +1,3 @@
-require('./class-api-stories');
+require('./modal-manager-stories');
 require('./context-api-stories');
 require('./component-stories');
